projectCount 타입 변경 사업자 번호 중복 시 DuplicatedValueException 발생 처리

+++ app/src/main/java/com/howoocast/hywtl_has/business/view/BusinessListView.java
@@ -15,7 +15,7 @@ public class BusinessListView {
     private String address; // 주소
     private String officePhone; // 대표 전화번호
     private Integer managerCount; // 담당자
-    private Long projectCount; // 참여 프로젝트 총 개수
+    private Integer projectCount; // 참여 프로젝트 총 개수
     private String memo;
 
     public static BusinessListView assemble(Business source) {
